fix: break refresh URL import cycle
