+++ src/server/auth.ts
@@ -6,9 +6,9 @@ import { SystemRoles } from 'librechat-data-provider';
 import { createServerFn } from '@tanstack/react-start';
 import { getRequestHeader } from '@tanstack/react-start/server';
 import type * as t from '@/types';
-import { useAppSession, SESSION_CONFIG } from './session';
-import { getApiBaseUrl, getServerApiUrl } from './utils/api';
+import { getApiBaseUrl, getServerApiUrl } from './utils/url';
 import { refreshAdminTokenDeduped } from './utils/refresh';
+import { useAppSession, SESSION_CONFIG } from './session';
 
 /** Extract a named cookie value from `set-cookie` response headers. */
 function extractCookieValue(response: Response, name: string): string | undefined {
